Update README.md
Tiny typo for SSH (should be SSL) and minor layout amends.

This homebridge plugin does not work for 2018 on my 65OLED803/12. I can turn the TV on and off, but ambilight control and volume does not work, and trying to change input just briefly flashes the inout menu. While I can get to some JointSpace URLs, the /1/system and / URLs just shows a page with "Not Found". Where do I log issues?