Update README.md
Tiny typo for SSH (should be SSL), showed that version should be incremented in the JointSpace URL to detect the API version, and minor layout amends.

This homebridge plugin does not work for 2018 on my 65OLED803/12. I can turn the TV on and off, but ambilight control and volume does not work, and trying to change input just briefly flashes the inout menu. The TV is using JointSpace API 6.1. Where do I log issues?